refactor(payment-cli): split payment_cli monolith by responsibility
Break the 1274-line openfatture/payment/cli/payment_cli.py into a package:
_app (Typer app + shared helpers) + accounts/transactions/reconciliation/
reminders/reporting command modules. Command bodies are pure copy-paste.

Tests updated (no shim): mock.patch targets repointed from payment_cli.<symbol>
to the module owning each exercised command (e.g. reconciliation.ReconciliationService,
transactions.ImporterFactory), with command modules importing those symbols so
they remain patchable at the correct lookup site.

Verified: command tree identical (18 commands, golden diff empty),
tests/payment/cli/test_payment_cli.py 41/41 green, ruff clean.
